--- packages/remix/experimental/api.server.d.ts
@@ -1,1 +1,0 @@
-export * from '../dist/experimental/api';
