--- packages/remix/experimental/api.server.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/experimental/api/index');
