--- packages/remix/experimental/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/experimental/index');
