--- packages/remix/experimental/ssr.server.d.ts
@@ -1,1 +1,0 @@
-export * from '../dist/experimental/ssr';
