--- packages/remix/experimental/ssr.server.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/experimental/ssr/index');
