+++ packages/remix/src/api/index.ts
@@ -1,7 +1,7 @@
-if (!process.env.CLERK_API_KEY && !process.env.CLERK_SECRET_KEY) {
-  throw Error(
-    'The CLERK_API_KEY or CLERK_SECRET_KEY environment variable must be set to use imports from @clerk/remix/api.',
-  );
-}
+import { Clerk } from '@clerk/backend';
 
-export * from '@clerk/clerk-sdk-node';
+const createClerkClient = Clerk;
+
+export { createClerkClient };
+
+export * from '@clerk/backend';
