--- packages/remix/src/experimental/api/index.ts
@@ -1,7 +1,0 @@
-import { Clerk } from '@clerk/backend';
-
-const createClerkClient = Clerk;
-
-export { createClerkClient };
-
-export * from '@clerk/backend';
