--- packages/remix/src/experimental/client/Interstitial.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export function Interstitial({ html }: { html: string }) {
-  return <html dangerouslySetInnerHTML={{ __html: html }} />;
-}
