--- packages/remix/src/experimental/client/index.ts
@@ -1,4 +1,0 @@
-export * from './RemixClerkProvider';
-export { ClerkApp } from './ClerkApp';
-export * from './ClerkCatchBoundary';
-export { WithClerkState } from './types';
