--- packages/remix/src/experimental/client/types.ts
@@ -1,16 +1,0 @@
-import { InitialState } from '@clerk/types';
-
-export type ClerkState = {
-  __type: 'clerkState';
-  __internal_clerk_state: {
-    __clerk_ssr_interstitial: string;
-    __clerk_ssr_state: InitialState;
-    __frontendApi: string;
-    __clerk_debug: any;
-  };
-};
-
-export type WithClerkState<U = any> = {
-  data: U;
-  clerkState: { __type: 'clerkState' };
-};
