--- packages/remix/src/experimental/globalPolyfill.ts
@@ -1,3 +1,0 @@
-if (typeof window !== 'undefined' && typeof (window as any).global === 'undefined') {
-  (window as any).global = window;
-}
