--- packages/remix/src/experimental/index.ts
@@ -1,3 +1,0 @@
-import './globalPolyfill';
-
-export * from './client';
