--- packages/remix/src/experimental/ssr/index.ts
@@ -1,2 +1,0 @@
-export * from './rootAuthLoader';
-export * from './getAuth';
