+++ packages/remix/src/ssr/authenticateRequest.ts
